bug affichage axe < 255

+++ Inc/menu.h
@@ -9,7 +9,7 @@
 
 uint8_t add_elem_menu(menu* my_menu, var* my_var);
 void fill_one_line_menu(uint8_t* str, var* my_var);
-void display_menu(menu *my_menu, uint16_t x_pos, uint8_t y_pos);
+void display_menu(menu *my_menu, uint16_t x_pos, uint16_t y_pos);
 void add_line_number(uint8_t* str, uint8_t position);
 void fill_string_with_space(uint8_t* str, uint8_t size);
 void menu_up(menu *my_menu);
